typecast ULL to uint64_t for ros::Time
git-svn-id: https://code.ros.org/svn/ros-pkg/pkg/trunk/visualization@6451 eb33c2ac-9c88-4c90-87e0-44a10359b0c3

+++ ogre_visualizer/src/ogre_visualizer/displays/map_display.cpp
@@ -324,7 +324,7 @@ void MapDisplay::load()
 
 void MapDisplay::transformMap()
 {
-  tf::Stamped<tf::Pose> pose( btTransform( btQuaternion( 0, 0, 0 ), btVector3( 0, 0, 0 ) ), ros::Time(0ULL), "map" );
+  tf::Stamped<tf::Pose> pose( btTransform( btQuaternion( 0, 0, 0 ), btVector3( 0, 0, 0 ) ), ros::Time((uint64_t)0ULL), "map" );
 
   try
   {
