+++ ogre_visualizer/src/ogre_visualizer/displays/poly_line_2d_display.cpp
@@ -276,7 +276,7 @@ void PolyLine2DDisplay::processMessage()
   clear();
 
   tf::Stamped<tf::Pose> pose( btTransform( btQuaternion( 0.0f, 0.0f, 0.0f ), btVector3( 0.0f, 0.0f, z_position_ ) ),
-                                ros::Time(0ULL), "map" );
+                                ros::Time((uint64_t)0ULL), "map" );
 
   try
   {
